+++ MARU/MARU/Screens/Library/PastMeetingViewController.swift
@@ -74,6 +74,11 @@ extension PastMeetingViewController: UICollectionViewDataSource {
       cell.rx.didTapEvaluateButton
         .subscribe(onNext: {
           // 방장이 아닌 경우에는 방장 평가하도록 처리
+          let viewController = EvaluateViewController()
+          viewController.modalTransitionStyle = .crossDissolve
+          viewController.modalPresentationStyle = .overCurrentContext
+          // currentContext로 하면 배경 투명효과 안들어감. 작동원리는 동일
+          self.present(viewController, animated: false, completion: nil)
         })
         .disposed(by: cell.disposeBag)
     }
